+++ test/unit/widgets/world/RenderSettingsWidgetTest.cpp
@@ -72,6 +72,7 @@ namespace RenderSettingsWidgetTest {
     EXPECT_EQ(256, widget.shadowMapSize());
     EXPECT_DOUBLE_EQ(0.001, widget.shadowBias());
     EXPECT_EQ(0, widget.shadowFilterRadius());
+    EXPECT_EQ(QString("PCF"), widget.shadowFilterMode());
   }
 
   TEST_F(RenderSettingsWidgetTest, ShouldDefaultRaytracerToPeriodicProgressDisplay) {
@@ -119,20 +120,24 @@ namespace RenderSettingsWidgetTest {
     auto shadowMapSize = widget.findChild<QSpinBox*>("rasterShadowMapSize");
     auto shadowBias = widget.findChild<QDoubleSpinBox*>("rasterShadowBias");
     auto filterRadius = widget.findChild<QSpinBox*>("rasterShadowFilterRadius");
+    auto filterMode = widget.findChild<QComboBox*>("rasterShadowFilterMode");
     ASSERT_NE(nullptr, shadowMaps);
     ASSERT_NE(nullptr, shadowMapSize);
     ASSERT_NE(nullptr, shadowBias);
     ASSERT_NE(nullptr, filterRadius);
+    ASSERT_NE(nullptr, filterMode);
 
     shadowMaps->setChecked(true);
     shadowMapSize->setValue(1024);
     shadowBias->setValue(0.25);
     filterRadius->setValue(3);
+    filterMode->setCurrentText("PCSS");
 
     EXPECT_TRUE(widget.shadowMapsEnabled());
     EXPECT_EQ(1024, widget.shadowMapSize());
     EXPECT_DOUBLE_EQ(0.25, widget.shadowBias());
     EXPECT_EQ(3, widget.shadowFilterRadius());
+    EXPECT_EQ(QString("PCSS"), widget.shadowFilterMode());
   }
 
   TEST_F(RenderSettingsWidgetTest, ShouldReadRasterPostAAControl) {
@@ -179,21 +184,25 @@ namespace RenderSettingsWidgetTest {
     auto shadowMapSize = widget.findChild<QSpinBox*>("rasterShadowMapSize");
     auto shadowBias = widget.findChild<QDoubleSpinBox*>("rasterShadowBias");
     auto filterRadius = widget.findChild<QSpinBox*>("rasterShadowFilterRadius");
+    auto filterMode = widget.findChild<QComboBox*>("rasterShadowFilterMode");
     ASSERT_NE(nullptr, engineType);
     ASSERT_NE(nullptr, shadowMaps);
     ASSERT_NE(nullptr, shadowMapSize);
     ASSERT_NE(nullptr, shadowBias);
     ASSERT_NE(nullptr, filterRadius);
+    ASSERT_NE(nullptr, filterMode);
 
     engineType->setCurrentText("Rasterizer");
     EXPECT_TRUE(shadowMapSize->isHidden());
     EXPECT_TRUE(shadowBias->isHidden());
     EXPECT_TRUE(filterRadius->isHidden());
+    EXPECT_TRUE(filterMode->isHidden());
 
     shadowMaps->setChecked(true);
     EXPECT_FALSE(shadowMapSize->isHidden());
     EXPECT_FALSE(shadowBias->isHidden());
     EXPECT_FALSE(filterRadius->isHidden());
+    EXPECT_FALSE(filterMode->isHidden());
   }
 
   TEST_F(RenderSettingsWidgetTest, ShouldReturnNonZeroResolution) {